requirements.txt
alembic==1.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
croniter==0.3.34
dnspython==1.16.0
email-validator==1.1.1
Faker==4.1.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-RQ2==18.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
natsort==7.0.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
rq==1.5.2
rq-scheduler==0.10.0
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.1
